+++ example/android/app/src/main/kotlin/com/example/appmetrica_sdk_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.example.appmetrica_sdk_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()