+++ lib/appmetrica_sdk.dart
@@ -5,7 +5,6 @@
 library appmetrica_sdk;
 
 import 'dart:async';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 part 'src/appmetrica_sdk.dart';