+++ lib/src/appmetrica_sdk.dart
@@ -13,10 +13,12 @@ class AppmetricaSdk {
   factory AppmetricaSdk() {
     if (_instance == null) {
       _instance = AppmetricaSdk.private(
-          const MethodChannel('emallstudio.com/appmetrica_sdk'));
+        const MethodChannel('emallstudio.com/appmetrica_sdk'),
+      );
     }
     return _instance!;
   }
+
   AppmetricaSdk.private(this._channel);
 
   /// Initializes the library in an application with given parameters.
@@ -28,13 +30,14 @@ class AppmetricaSdk {
   /// The [statisticsSending] enables/disables sending statistics to the AppMetrica server. By default, sending is enabled.
   /// The [crashReporting] indicating that sending app crashes is enabled. By default, sending is enabled.
   /// Android only. The [maxReportsInDatabaseCount] is maximum number of events that can be stored in the database on the phone before being sent to AppMetrica. If there are more events, old records will begin to be deleted. The default value is 1000 (allowed values from 100 to 10000).
-  Future<void> activate(
-      {required String apiKey,
-      int sessionTimeout = 10,
-      bool locationTracking = true,
-      bool statisticsSending = true,
-      bool crashReporting = true,
-      int maxReportsInDatabaseCount = 1000}) async {
+  Future<void> activate({
+    required String apiKey,
+    int sessionTimeout = 10,
+    bool locationTracking = true,
+    bool statisticsSending = true,
+    bool crashReporting = true,
+    int maxReportsInDatabaseCount = 1000,
+  }) async {
     await _channel.invokeMethod<void>('activate', <String, dynamic>{
       'apiKey': apiKey,
       'sessionTimeout': sessionTimeout,
@@ -44,13 +47,15 @@ class AppmetricaSdk {
       'maxReportsInDatabaseCount': maxReportsInDatabaseCount,
     });
 
-    /// Set the API Key after activation.
+    // Set the API Key after activation
     _apiKey = apiKey;
   }
 
   /// Sends an event with [name] and message as a set of [attributes] to the AppMetrica server.
-  Future<void> reportEvent(
-      {required String name, Map<String, dynamic>? attributes}) async {
+  Future<void> reportEvent({
+    required String name,
+    Map<String, dynamic>? attributes,
+  }) async {
     if (_apiKey == null) {
       throw 'The API key is not set';
     }
@@ -61,89 +66,113 @@ class AppmetricaSdk {
   }
 
   /// Sends custom string user profile attribute with the given [key] and [value] to the AppMetrica server.
-  Future<void> reportUserProfileCustomString(
-      {required String key, String? value}) async {
-    if (_apiKey == null) {
-      throw 'The API key is not set';
-    }
-    await _channel
-        .invokeMethod<void>('reportUserProfileCustomString', <String, dynamic>{
-      'key': key,
-      'value': value,
-    });
-  }
-
-  /// Sends custom number user profile attribute with the given [key] and [value] to the AppMetrica server.
-  Future<void> reportUserProfileCustomNumber(
-      {required String key, double? value}) async {
-    if (_apiKey == null) {
-      throw 'The API key is not set';
-    }
-    await _channel
-        .invokeMethod<void>('reportUserProfileCustomNumber', <String, dynamic>{
-      'key': key,
-      'value': value,
-    });
-  }
-
-  /// Sends custom boolean user profile attribute with the given [key] and [value] to the AppMetrica server.
-  Future<void> reportUserProfileCustomBoolean(
-      {required String key, bool? value}) async {
-    if (_apiKey == null) {
-      throw 'The API key is not set';
-    }
-    await _channel
-        .invokeMethod<void>('reportUserProfileCustomBoolean', <String, dynamic>{
-      'key': key,
-      'value': value,
-    });
-  }
-
-  /// Sends custom counter user profile attribute with the given [key] and [value] to the AppMetrica server.
-  Future<void> reportUserProfileCustomCounter(
-      {required String key, required double delta}) async {
-    if (_apiKey == null) {
-      throw 'The API key is not set';
-    }
-    await _channel
-        .invokeMethod<void>('reportUserProfileCustomCounter', <String, dynamic>{
-      'key': key,
-      'delta': delta,
-    });
-  }
-
-  /// Sends predefined [userName] profile attribute to the AppMetrica server.
-  Future<void> reportUserProfileUserName({String? userName}) async {
-    if (_apiKey == null) {
-      throw 'The API key is not set';
-    }
-    await _channel
-        .invokeMethod<void>('reportUserProfileUserName', <String, dynamic>{
-      'userName': userName,
-    });
-  }
-
-  /// Sends predefined [notificationsEnabled] profile attribute to the AppMetrica server.
-  Future<void> reportUserProfileNotificationsEnabled(
-      {bool? notificationsEnabled}) async {
+  Future<void> reportUserProfileCustomString({
+    required String key,
+    String? value,
+  }) async {
     if (_apiKey == null) {
       throw 'The API key is not set';
     }
     await _channel.invokeMethod<void>(
-        'reportUserProfileNotificationsEnabled', <String, dynamic>{
-      'notificationsEnabled': notificationsEnabled,
-    });
+      'reportUserProfileCustomString',
+      <String, dynamic>{
+        'key': key,
+        'value': value,
+      },
+    );
+  }
+
+  /// Sends custom number user profile attribute with the given [key] and [value] to the AppMetrica server.
+  Future<void> reportUserProfileCustomNumber({
+    required String key,
+    double? value,
+  }) async {
+    if (_apiKey == null) {
+      throw 'The API key is not set';
+    }
+    await _channel.invokeMethod<void>(
+      'reportUserProfileCustomNumber',
+      <String, dynamic>{
+        'key': key,
+        'value': value,
+      },
+    );
+  }
+
+  /// Sends custom boolean user profile attribute with the given [key] and [value] to the AppMetrica server.
+  Future<void> reportUserProfileCustomBoolean({
+    required String key,
+    bool? value,
+  }) async {
+    if (_apiKey == null) {
+      throw 'The API key is not set';
+    }
+    await _channel.invokeMethod<void>(
+      'reportUserProfileCustomBoolean',
+      <String, dynamic>{
+        'key': key,
+        'value': value,
+      },
+    );
+  }
+
+  /// Sends custom counter user profile attribute with the given [key] and [value] to the AppMetrica server.
+  Future<void> reportUserProfileCustomCounter({
+    required String key,
+    required double delta,
+  }) async {
+    if (_apiKey == null) {
+      throw 'The API key is not set';
+    }
+    await _channel.invokeMethod<void>(
+      'reportUserProfileCustomCounter',
+      <String, dynamic>{
+        'key': key,
+        'delta': delta,
+      },
+    );
+  }
+
+  /// Sends predefined [userName] profile attribute to the AppMetrica server.
+  Future<void> reportUserProfileUserName([String? userName]) async {
+    if (_apiKey == null) {
+      throw 'The API key is not set';
+    }
+    await _channel.invokeMethod<void>(
+      'reportUserProfileUserName',
+      <String, dynamic>{
+        'userName': userName,
+      },
+    );
+  }
+
+  /// Sends predefined [notificationsEnabled] profile attribute to the AppMetrica server.
+  Future<void> reportUserProfileNotificationsEnabled({
+    bool? notificationsEnabled,
+  }) async {
+    if (_apiKey == null) {
+      throw 'The API key is not set';
+    }
+    await _channel.invokeMethod<void>(
+      'reportUserProfileNotificationsEnabled',
+      <String, dynamic>{
+        'notificationsEnabled': notificationsEnabled,
+      },
+    );
   }
 
   /// Disable and enable sending statistics to the AppMetrica server.
-  Future<void> setStatisticsSending({required bool statisticsSending}) async {
+  Future<void> setStatisticsSending(bool statisticsSending) async {
     if (_apiKey == null) {
       throw 'The API key is not set';
     }
 
-    await _channel.invokeMethod<void>('setStatisticsSending', <String, dynamic>{
-      'statisticsSending': statisticsSending,
-    });
+    await _channel.invokeMethod<void>(
+      'setStatisticsSending',
+      <String, dynamic>{
+        'statisticsSending': statisticsSending,
+      },
+    );
   }
 
   /// Returns the current version of the AppMetrica library.
@@ -156,13 +185,16 @@ class AppmetricaSdk {
 
   /// Sets the ID of the user profile.
   /// Required for predefined profile attributes like Name or Notifications enabled.
-  Future<void> setUserProfileID({required String userProfileID}) async {
+  Future<void> setUserProfileID(String? userProfileID) async {
     if (_apiKey == null) {
       throw 'The API key is not set';
     }
-    await _channel.invokeMethod<void>('setUserProfileID', <String, dynamic>{
-      'userProfileID': userProfileID,
-    });
+    await _channel.invokeMethod<void>(
+      'setUserProfileID',
+      <String, dynamic>{
+        'userProfileID': userProfileID,
+      },
+    );
   }
 
   /// Sends stored events from the buffer.
@@ -182,13 +214,16 @@ class AppmetricaSdk {
   // some specific traffic sources like Facebook.
   // Referral URL reporting is no longer available
   // but many agencies use this report
-  Future<void> reportReferralUrl({required String referral}) async {
+  Future<void> reportReferralUrl(String referral) async {
     if (_apiKey == null) {
       throw 'The API key is not set';
     }
-    await _channel.invokeMethod<String>('reportReferralUrl', <String, dynamic>{
-      'referral': referral,
-    });
+    await _channel.invokeMethod<String>(
+      'reportReferralUrl',
+      <String, dynamic>{
+        'referral': referral,
+      },
+    );
     return;
   }
 }